[Release] v1.0.2 (#18)
* [Feature] CI/CD Workflow 생성 (#7)

* fix: Watchdog Interval 코드 반영

* fix: type: "none" 케이스 제거

* chore: CI workflow

* chore: CD release workflow

* docs: PR Template 생성

* [Fix] 볼륨이 0일 때 재생이 중단되는 문제 해결 (#9)

* fix: background throttling 해제하기

* docs: 배경 음악 재생이 중단되는 문제 정리

* [Feature] 버전 업데이트 및 앱 삭제 설정 (#10)

* chore: electron updater 의존성

* chore: 배포 시 버전 확인용 yml 파일 첨부

* feat: electron-updater 연동

* feat: IPC 브릿지

* feat: 업데이트 상태 관리

* feat: 업데이트 알림 UI

* fix: debug용 코드 제거

* docs: 자동 업데이트 과정 정리

* chore: 삭제 시 AppData도 삭제 가능

* fix: 업데이트 체크 중 에러는 감추기

* refactor: 파일 의미에 맞게 파일명 변경

* chore: v1.0.0 버전 업데이트 (#12)

* chore: main 브랜치 변경사항 develop에 반영 (#16)

* [Release] v1.0.0 (#11)

* fix: 갑자기 배경 음악이 재생되지 않는 버그 수정 (#6)

* refactor: 실제 시간과의 오차 감소 시키기

* fix: 오디오 세션 중단 문제 해결

* docs: 오디오 세션 중단 문제 해결 과정

* [Feature] CI/CD Workflow 생성 (#7)

* fix: Watchdog Interval 코드 반영

* fix: type: "none" 케이스 제거

* chore: CI workflow

* chore: CD release workflow

* docs: PR Template 생성

* [Fix] 볼륨이 0일 때 재생이 중단되는 문제 해결 (#9)

* fix: background throttling 해제하기

* docs: 배경 음악 재생이 중단되는 문제 정리

* [Feature] 버전 업데이트 및 앱 삭제 설정 (#10)

* chore: electron updater 의존성

* chore: 배포 시 버전 확인용 yml 파일 첨부

* feat: electron-updater 연동

* feat: IPC 브릿지

* feat: 업데이트 상태 관리

* feat: 업데이트 알림 UI

* fix: debug용 코드 제거

* docs: 자동 업데이트 과정 정리

* chore: 삭제 시 AppData도 삭제 가능

* fix: 업데이트 체크 중 에러는 감추기

* refactor: 파일 의미에 맞게 파일명 변경

* chore: v1.0.0 버전 업데이트

* [Hotfix] 업데이트 시에도 AppData 삭제 불가능 (#14)

* fix: build-debug.yml 파일 업로드 제외

* fix: 업데이트 시엔 AppData 삭제 여부 묻지 않도록 하기

* chore: v1.0.1 버전 설정

* docs: 업로드 시 AppData 삭제 문제 정리

* fix: 원인 분석 수정

* [Hotfix] hotfix 브랜치 CD 워크플로우 트리거 추가 (#15)

* fix: build-debug.yml 파일 업로드 제외

* fix: 업데이트 시엔 AppData 삭제 여부 묻지 않도록 하기

* chore: v1.0.1 버전 설정

* docs: 업로드 시 AppData 삭제 문제 정리

* fix: 원인 분석 수정

* fix: hotfix 브랜치일 때도 main에 merge하고 release 하기

* [Refactor] 배경 음악 코드 리팩터링 (#17)

* fix: AppData 삭제 조건 수정

* chore: src `@` 별칭 설정

* chore: modelNameMapper 추가

* refactor: import 경로 변경

* refactor: 디렉토리 구조 변경

* test: 재생 관련 기능 테스트

* refactor: src/type을 각 features로 분리

* refactor: useBgMusic.ts 코드 분리

* refactor: features 단위로 디렉토리 구조 변경

* docs: features 단위로 코드 분리한 과정

* refactor: import 경로 재설정

* fix: defaultSchedule 데이터 추가

* release: v1.0.2
