+++ jest.config.cjs
@@ -5,6 +5,7 @@ module.exports = {
     "^.+\\.tsx?$": ["ts-jest", { tsconfig: "./tsconfig.test.json" }],
   },
   moduleNameMapper: {
+    "^@/(.*)$": "<rootDir>/src/$1",
     "\\.(css|scss)$": "<rootDir>/src/__mocks__/fileMock.ts",
   },
   setupFilesAfterEnv: ["<rootDir>/src/setupTests.ts"],
