+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "gogo-dino-annc",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "private": true,
   "main": "public/electron.js",
   "scripts": {
