+++ src/features/announcement/components/AnnouncementCard/AnnouncementCard.scss
@@ -1,4 +1,4 @@
-@use "../../styles/variables" as *;
+@use "../../../../styles/variables" as *;
 
 .announcement-card {
   background: #ffffff;
