+++ src/features/announcement/components/CategorySection/CategorySection.scss
@@ -1,4 +1,4 @@
-@use "../../styles/variables" as *;
+@use "../../../../styles/variables" as *;
 
 .category-section {
   display: flex;
