+++ src/features/announcement/components/ScheduleSettings/ScheduleSettings.scss
@@ -1,4 +1,4 @@
-@use "../../styles/variables" as *;
+@use "../../../../styles/variables" as *;
 
 .modal-overlay {
   position: fixed;
