+++ src/features/announcement/types/category.ts
@@ -1,0 +1,1 @@
+export type Category = "attraction" | "closing" | "table";
