+++ src/features/bg-music/components/BgMusicPanel/BgMusicPanel.scss
@@ -1,4 +1,4 @@
-@use "../../styles/variables" as *;
+@use "../../../../styles/variables" as *;
 
 .bg-music-panel {
   display: flex;
