--- src/types/category.ts
@@ -1,1 +1,0 @@
-export type Category = "attraction" | "closing" | "table";
